+++ app/src/main/java/com/hacka/presenteperfeito/core/designSystem/components/buttons/ProjectButtonType.kt
@@ -1,0 +1,6 @@
+package com.hacka.presenteperfeito.core.designSystem.components.buttons
+
+sealed interface ProjectButtonTypes {
+    data object Primary : ProjectButtonTypes
+    data object Secondary : ProjectButtonTypes
+}
